# sql/schemachanger/scplan/internal/opgen: optimize allocations