--- a/pkg/sql/schemachanger/scplan/internal/opgen/opgen_enum_type.go
+++ b/pkg/sql/schemachanger/scplan/internal/opgen/opgen_enum_type.go
@@ -47,7 +47,7 @@ func init() {
 			),
 			to(scpb.Status_ABSENT,
 				minPhase(scop.PostCommitPhase),
-				emit(func(this *scpb.EnumType, md targetsWithElementMap) scop.Op {
+				emit(func(this *scpb.EnumType, md *targetsWithElementMap) scop.Op {
 					return newLogEventOp(this, md)
 				}),
 				emit(func(this *scpb.EnumType) scop.Op {