--- a/pkg/sql/schemachanger/scplan/internal/opgen/opgen_primary_index.go
+++ b/pkg/sql/schemachanger/scplan/internal/opgen/opgen_primary_index.go
@@ -89,7 +89,7 @@ func init() {
 				}),
 			),
 			to(scpb.Status_ABSENT,
-				emit(func(this *scpb.PrimaryIndex, md targetsWithElementMap) scop.Op {
+				emit(func(this *scpb.PrimaryIndex, md *targetsWithElementMap) scop.Op {
 					return &scop.CreateGcJobForIndex{
 						TableID:             this.TableID,
 						IndexID:             this.IndexID,