--- a/pkg/sql/schemachanger/scplan/internal/opgen/opgen_schema.go
+++ b/pkg/sql/schemachanger/scplan/internal/opgen/opgen_schema.go
@@ -46,7 +46,7 @@ func init() {
 			),
 			to(scpb.Status_ABSENT,
 				minPhase(scop.PostCommitPhase),
-				emit(func(this *scpb.Schema, md targetsWithElementMap) scop.Op {
+				emit(func(this *scpb.Schema, md *targetsWithElementMap) scop.Op {
 					return newLogEventOp(this, md)
 				}),
 				emit(func(this *scpb.Schema) scop.Op {