--- a/pkg/sql/schemachanger/scplan/internal/opgen/opgen_sequence.go
+++ b/pkg/sql/schemachanger/scplan/internal/opgen/opgen_sequence.go
@@ -52,10 +52,10 @@ func init() {
 			),
 			to(scpb.Status_ABSENT,
 				minPhase(scop.PostCommitPhase),
-				emit(func(this *scpb.Sequence, md targetsWithElementMap) scop.Op {
+				emit(func(this *scpb.Sequence, md *targetsWithElementMap) scop.Op {
 					return newLogEventOp(this, md)
 				}),
-				emit(func(this *scpb.Sequence, md targetsWithElementMap) scop.Op {
+				emit(func(this *scpb.Sequence, md *targetsWithElementMap) scop.Op {
 					return &scop.CreateGcJobForTable{
 						TableID:             this.SequenceID,
 						StatementForDropJob: statementForDropJob(this, md),